# docs/swagger.yaml
basePath: /[v1]/
definitions:
...
  github_com_semenovem_portal_pkg_fail.Response:
    properties:
      additional_fields:
        additionalProperties: true
        type: object
      code:
        type: integer
      message:
        type: string
      validation_errors:
        items:
          $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.ValidationError'
        type: array
    type: object
  github_com_semenovem_portal_pkg_fail.ValidationError:
    properties:
      message:
        type: string
      path:
        type: string
    type: object
  internal_abc_auth_controller.loginForm:
    properties:
      device_id:
        type: string
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  internal_abc_auth_controller.loginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        description: TODO для разработки
        type: string
      user_id:
        type: integer
    type: object
  internal_abc_auth_controller.onetimeAuthForm:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
  internal_abc_auth_controller.onetimeAuthResponse:
    properties:
      entry_id:
        type: string
      uri:
        type: string
    type: object
...
  internal_abc_media_controller.fileUploadResponse:
    properties:
      file:
        $ref: '#/definitions/internal_abc_media_controller.fileUploadView'
    type: object
  internal_abc_media_controller.fileUploadView:
    properties:
      id:
        type: integer
      preview_link:
        description: uri
        type: string
    type: object
...
  internal_abc_people_controller.userPublicProfileView:
    properties:
      avatar:
        type: string
      firstname:
        type: string
      id:
        type: integer
      surname:
        type: string
    type: object
  internal_abc_store_controller.loadView:
    properties:
      payload:
        type: string
    type: object
  internal_abc_store_controller.storeForm:
    properties:
      payload:
        type: string
      storePath:
        type: string
    required:
    - payload
    - storePath
    type: object
...
host: localhost:8080
info:
  contact:
    email: semenovem@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: portal API
  version: "1.0"
paths:
  /auth/login:
    post:
      parameters:
      - description: Логин/пароль
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_abc_auth_controller.loginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_auth_controller.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/logout:
    post:
      parameters:
      - description: refresh токен
        in: header
        name: refresh-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_auth_controller.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Выход из авторизованной сессии пользователя
      tags:
      - auth
  /auth/onetime:
    post:
      parameters:
      - description: данные для создания сессии
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_abc_auth_controller.onetimeAuthForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_auth_controller.onetimeAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Создание ссылки для одноразовой авторизации
      tags:
      - auth
  /auth/onetime/:entry_id:
    post:
      parameters:
      - description: id сессии авторизации
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_auth_controller.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Логин по одноразовой ссылке
      tags:
      - auth
...
  /media/file:
    post:
      consumes:
      - multipart/form-data
      description: |
        note - подпись к файлу (опционально)
        file - файл
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_abc_media_controller.fileUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Сохранение файлов
      tags:
      - media
...
  /people/:user_id/profile:
    get:
      description: |
        Проверяет действующие права на просмотр расширенных данных пользователя
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_people_controller.userPublicProfileView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя по его ID
      tags:
      - people
  /people/:user_id/profile/public:
    get:
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_people_controller.userPublicProfileView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Получить публичный профиль пользователя по его ID
      tags:
      - people
...
  /people/self/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_people_controller.userPublicProfileView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Получить свой профиль
      tags:
      - people
  /store/:store_path:
    delete:
      parameters:
      - description: code store
        in: path
        name: store_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Удаление
      tags:
      - store
    get:
      parameters:
      - description: code store
        in: path
        name: store_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_abc_store_controller.loadView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Чтение произвольных клиентских данных
      tags:
      - store
    post:
      description: |
        Для возможности восстановления состоянии на клиенте
      parameters:
      - description: code store
        in: path
        name: store_code
        required: true
        type: string
      - description: Данные для сохранения
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_abc_store_controller.storeForm'
      produces:
      - application/json
      responses:
        "201":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_semenovem_portal_pkg_fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Сохранение произвольных клиентских данных
      tags:
      - store
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"